Use float for num1's conversion also

+++ calc.py
@@ -6,7 +6,7 @@ def compute(expression):
         '1 - 2' -> -1
     """
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = float(num0), int(num1)
+    num0, num1 = float(num0), float(num1)
 
     if operator == '+':
         return num0 + num1
